fix: num_macro_particles is int

+++ libpic/patch/patch.py
@@ -368,7 +368,7 @@ class Patches:
                 species.ppc,
             )
         else:
-            num_macro_particles = np.zeros(self.npatches)
+            num_macro_particles = np.zeros(self.npatches, dtype=int)
 
 
         for ipatch in range(self.npatches):
